+++ lib/Server.js
@@ -39,7 +39,12 @@ class Server {
     // this value of web socket can be overwritten for tests
     this.webSocketHeartbeatInterval = 30000;
 
-    normalizeOptions(this.compiler, this.options, this.logger);
+    normalizeOptions(
+      this.compiler,
+      this.options,
+      this.logger,
+      Server.findCacheDir()
+    );
 
     this.applyDevServerPlugin();
