+++ lib/utils/normalizeOptions.js
@@ -1,14 +1,15 @@
 'use strict';
 
+const os = require('os');
 const path = require('path');
+const del = require('del');
 const fs = require('graceful-fs');
-const isAbsoluteUrl = require('is-absolute-url');
 const getCompilerConfigArray = require('./getCompilerConfigArray');
 
-function normalizeOptions(compiler, options, logger) {
+function normalizeOptions(compiler, options, logger, cacheDir) {
   // TODO: improve this to not use .find for compiler watchOptions
-  const configArr = getCompilerConfigArray(compiler);
-  const watchOptionsConfig = configArr.find(
+  const configArray = getCompilerConfigArray(compiler);
+  const watchOptionsConfig = configArray.find(
     (config) => config.watch !== false && config.watchOptions
   );
   const watchOptions = watchOptionsConfig
@@ -45,6 +46,8 @@ function normalizeOptions(compiler, options, logger) {
   }
 
   if (options.static) {
+    const isAbsoluteUrl = require('is-absolute-url');
+
     options.static.forEach((staticOption) => {
       if (isAbsoluteUrl(staticOption.directory)) {
         throw new Error('Using a URL as static.directory is not supported');
@@ -180,8 +183,6 @@ function normalizeOptions(compiler, options, logger) {
 
   // https option
   if (options.https) {
-    const getCertificate = require('./getCertificate');
-
     for (const property of ['cacert', 'pfx', 'key', 'cert']) {
       const value = options.https[property];
       const isBuffer = value instanceof Buffer;
@@ -205,7 +206,100 @@ function normalizeOptions(compiler, options, logger) {
     let fakeCert;
 
     if (!options.https.key || !options.https.cert) {
-      fakeCert = getCertificate(logger);
+      const certificateDir = cacheDir || os.tmpdir();
+      const certificatePath = path.join(certificateDir, 'server.pem');
+      let certificateExists = fs.existsSync(certificatePath);
+
+      if (certificateExists) {
+        const certificateTtl = 1000 * 60 * 60 * 24;
+        const certificateStat = fs.statSync(certificatePath);
+
+        const now = new Date();
+
+        // cert is more than 30 days old, kill it with fire
+        if ((now - certificateStat.ctime) / certificateTtl > 30) {
+          logger.info('SSL Certificate is more than 30 days old. Removing.');
+
+          del.sync([certificatePath], { force: true });
+
+          certificateExists = false;
+        }
+      }
+
+      if (!certificateExists) {
+        logger.info('Generating SSL Certificate');
+
+        const selfsigned = require('selfsigned');
+        const attributes = [{ name: 'commonName', value: 'localhost' }];
+        const pems = selfsigned.generate(attributes, {
+          algorithm: 'sha256',
+          days: 30,
+          keySize: 2048,
+          extensions: [
+            // {
+            //   name: 'basicConstraints',
+            //   cA: true,
+            // },
+            {
+              name: 'keyUsage',
+              keyCertSign: true,
+              digitalSignature: true,
+              nonRepudiation: true,
+              keyEncipherment: true,
+              dataEncipherment: true,
+            },
+            {
+              name: 'extKeyUsage',
+              serverAuth: true,
+              clientAuth: true,
+              codeSigning: true,
+              timeStamping: true,
+            },
+            {
+              name: 'subjectAltName',
+              altNames: [
+                {
+                  // type 2 is DNS
+                  type: 2,
+                  value: 'localhost',
+                },
+                {
+                  type: 2,
+                  value: 'localhost.localdomain',
+                },
+                {
+                  type: 2,
+                  value: 'lvh.me',
+                },
+                {
+                  type: 2,
+                  value: '*.lvh.me',
+                },
+                {
+                  type: 2,
+                  value: '[::1]',
+                },
+                {
+                  // type 7 is IP
+                  type: 7,
+                  ip: '127.0.0.1',
+                },
+                {
+                  type: 7,
+                  ip: 'fe80::1',
+                },
+              ],
+            },
+          ],
+        });
+
+        fs.mkdirSync(certificateDir, { recursive: true });
+        fs.writeFileSync(certificatePath, pems.private + pems.cert, {
+          encoding: 'utf8',
+        });
+      }
+
+      fakeCert = fs.readFileSync(certificatePath);
     }
 
     options.https.key = options.https.key || fakeCert;
