+++ test/fixtures/cli-colors-default-stats/webpack.config.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const ExitOnDonePlugin = require('../../helpers/ExitOnDonePlugin');
+
+module.exports = {
+  mode: 'development',
+  context: __dirname,
+  entry: './foo.js',
+  plugins: [ExitOnDonePlugin],
+};
