+++ test/fixtures/cli-colors-disabled/foo.js
@@ -1,0 +1,3 @@
+'use strict';
+
+console.log('i am foo!');
