+++ test/fixtures/cli-colors-disabled/webpack.config.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const ExitOnDonePlugin = require('../../helpers/ExitOnDonePlugin');
+
+module.exports = {
+  mode: 'development',
+  stats: {
+    colors: false,
+  },
+  context: __dirname,
+  entry: './foo.js',
+  plugins: [ExitOnDonePlugin],
+};
